First test/requirements.txt
google-generativeai
semanticscholar
feedparser
duckduckgo-search
